feat: Change v8go dependency
This changes v8go to use a fork maintained under the Gost-DOM project,
eliminating the need for custom installation steps.

That fork isn't intended to be be an independent v8go fork, and all
changes should eventually be merged back to tommie/v8go, eventually
eliminating the need for a dedicated fork.