refactor: Remove Event interface
The events were designed after an inheritance hierarchy in the browser
object model. But this didn't seem to that particularly helpful on the
Go side.

All differences in the inherited Event subtypes are represented by the
initialization data passed to the event constructor. Having that as a
property on the event seems sufficient.