refactor: Introduce a BrowsingContext interface
This allows some types to depend on a context, but not the entire Window
API, e.g., XHR can receive a window object containing current URL (used
to interpret local relative/absolute URLs), as well as the http client,
and logger to use going forward.

The clock is still an explicit argument, as the XHR need to have more
low level access to the clock than we want to expose on the Window
interface.