feat: Close browser when context is closed
If a context.Context was passed to the browser, the browser will
automatically close when the context cancels, possibly permitting script
engines to be recycled.

With a sensible amount of test-helpers to create a browser, accepting a
`testing.TB` as input, the browser can easily be configured to clean up
when the test is done.