fix: Don't write to logger when closing browser
When the browser is closed by a cancelled context, the logger may not be
valid anymore, so this code path should not log.

This also modifies the TestLogger to "close" in the test cleanup phase,
but this alone is not enough - that would merely get rid of the symptoms
in this test suite, but permit other test suites to experience the same
issue.