fix: Fix TestLogger to now write after test is done
This issue cannot be fixed in the browser itself, as it leads itself to
a race condition. By the time it is notified of the cancelled context,
the test may have completed yet.

So it must be up to the caller to ensire it passes a logger that doesn't
throw, even _after_ the test has completed.