test: create test verifying null signal/body doesn't throw
PR Review indicated that decoding `null` values for these two options
could result in an error. The test was meant to provoke the error and
drive a fix, but apparently the condition will be filtered elsewhere.

Code should be improved, so it is more obvious that this will not result
in an error when looking at the implementation directly; but for now,
the test is allowed to stay as a safe guard.